Switched to AVA as the testing framework and removed mocha and chai dependencies
