+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "fibtastic",
-  "version": "2.0.2",
+  "version": "2.0.3",
   "description": "Calculate a Fibonacci sequence up to a certain number of integers or to get just one specific value from anywhere in the sequence",
   "main": "index.js",
   "module": "fibtastic.js",
@@ -12,7 +12,7 @@
     "build:js": "babel src --out-dir lib --extensions \".ts,.tsx\" --source-maps inline",
     "bundle": "rollup -c",
     "bundle:watch": "rollup -c --watch",
-    "test": "rollup -c && mocha"
+    "test": "ava"
   },
   "keywords": [
     "fibonacci",
@@ -38,8 +38,7 @@
     "@babel/plugin-proposal-object-rest-spread": "^7.5.5",
     "@babel/preset-env": "^7.5.5",
     "@babel/preset-typescript": "^7.3.3",
-    "chai": "^4.2.0",
-    "mocha": "^6.2.0",
+    "ava": "^2.2.0",
     "rollup": "^1.17.0",
     "rollup-plugin-babel": "^4.3.3",
     "rollup-plugin-commonjs": "^10.0.1",
